requirements.txt
groq
SpeechRecognition
AppOpener
edge-tts
pygame
python-dotenv
PyAudio
spotipy
